Interim commit
Fixed a bug when c is calculated in con2vert(), the check to see if it's
in the region used to count points lying on a hyperplane as being in the
region, when really is should be inside the region and not on the
boundary. This caused divide by zero errors. The method is still not
perfect because the optimisation function still doesn't seem to find a
point inside the region. May need to change the method to find the
Chebyshev centre of the polytope instead.