Adjustment to in_region() and interior point check.
-Small wording fix to in_region() and out_region() docstrings.
-Fixed the check to see if c is truly in the region. Now we check if c
is out of the region or on the boundary. We can't in_region() because
in_region() will return True if it's on the boundary, which won't work
for our needs. May be useful to include an optional parameter to check
if boundary points are considered in the region in in_region().
-Also used a random point generation instead of the optimisation routine
just to see if it could work.